Изменение цены
Увеличение цены в меню улучшений + настраиваемая процентная прокачка, нужно исправить текст в меню улучшений, так как он не показывает увеличение цены

+++ Mig15/Assets/Script/Progress.cs
@@ -6,8 +6,8 @@ using UnityEngine;
 public class PlayerInfo
 {
     public int Coins = 0;
-    public int MaxHP = 5;
-    public int Damage = 2;
+    public float MaxHP = 5;
+    public float Damage = 2;
     public float TimeBetwinShots = 0.5f;
     public int Level = 0;
 }
